+++ CHANGELOG.rst
@@ -29,6 +29,7 @@ Bugfixes
 - Fixed an issue with wrong signal signaturs in WorkflowTreeEditManager.
 - Fixed a bug in the Remove1dPolynomialBackground plugin which forced a 
   polynomial order setting of 3.
+- Fixed an issue with pyFAI's calib2 app and additional argparse arguments.
 
 v23.07.05
 =========